requirements.txt
bottle==0.13.2
packaging==24.2
proxy_tools==0.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pywebview==5.3.2
QtPy==2.4.2
typing_extensions==4.12.2
